Seat count not hard coded anymore > class, services, required changes + db changes made

+++ src/main/java/com/example/CarGo/services/CarService.java
@@ -53,7 +53,7 @@ public class CarService {
     public List<Car> findCarsWithFilters(Location location,
                                          GearboxType gearboxType,
                                          ChassisType chassisType,
-                                         Integer seatCount,
+                                         Long seatCount,
                                          Integer yearMin,
                                          Integer yearMax,
                                          Double priceMin,
